splitting up userpane smart/dumb
