+++ shared/search/dumb.desktop.js
@@ -1,7 +1,11 @@
 // @flow
 
+import React from 'react'
 import Search from './render'
 import type {DumbComponentMap} from '../constants/types/more'
+import UserPane from './user-pane/render'
+
+import userPaneMocks from './user-pane/dumb.desktop'
 
 const results = [
   {
@@ -101,6 +105,7 @@ const commonUsers = [
 
 const commonProps = {
   username: 'bob',
+  userPane: <UserPane mode='keybase' userInfoProps={userPaneMocks['Search User Pane'].mocks['Unfollowed']} />,
   onSearch: text => console.log('OnSearch: ', text),
   searchHintText: 'Search Keybase',
   searchText: 'malg',
@@ -140,6 +145,25 @@ const searchMap: DumbComponentMap<Search> = {
     'Group non-user': {
       ...commonProps,
       showUserGroup: true,
+      userPane: <UserPane mode='external' nonUserInfoProps={userPaneMocks['Search Non-User Pane'].mocks['Normal']} />,
+      userForInfoPane: commonUsers[2],
+    },
+    'Group non-user No Avatar': {
+      ...commonProps,
+      showUserGroup: true,
+      userPane: <UserPane mode='external' nonUserInfoProps={userPaneMocks['Search Non-User Pane'].mocks['No Avatar']} />,
+      userForInfoPane: commonUsers[2],
+    },
+    'Group non-user Out of invites': {
+      ...commonProps,
+      showUserGroup: true,
+      userPane: <UserPane mode='external' nonUserInfoProps={userPaneMocks['Search Non-User Pane'].mocks['Out of invites']} />,
+      userForInfoPane: commonUsers[2],
+    },
+    'Group non-user Has Invite': {
+      ...commonProps,
+      showUserGroup: true,
+      userPane: <UserPane mode='external' nonUserInfoProps={userPaneMocks['Search Non-User Pane'].mocks['Has Invite']} />,
       userForInfoPane: commonUsers[2],
     },
     'Chat enabled': {
