+++ shared/search/user-pane/index.js
@@ -1,73 +1,22 @@
 // @flow
-import React, {Component} from 'react'
-import UserInfo from './user.render'
-import NonUserInfo from './non-user.render'
-import Help from './help'
-import Loading from './loading'
-
 import {fullName} from '../../constants/search'
 import keybaseUrl from '../../constants/urls'
 import {TypedConnector} from '../../util/typed-connect'
-
 import {getProfile, onFollow, onUnfollow} from '../../actions/tracker'
-
 import openURL from '../../util/open-url'
+import Render from './render'
 
-import type {Props as UserInfoPaneProps} from './user.render'
-import type {Props as NonUserInfoProps} from './non-user.render'
-
+import type {Props} from './render'
 import type {TypedState} from '../../constants/reducer'
 import type {SearchActions} from '../../constants/search'
 import type {TypedDispatch} from '../../constants/types/flux'
 
-type OwnProps = {
-  userInfoProps?: ?UserInfoPaneProps,
-}
-
-type Props = {
-  mode: 'keybase',
-  userInfoProps: UserInfoPaneProps,
-} | {
-  mode: 'external',
-  nonUserInfoProps: NonUserInfoProps,
-} | {
-  mode: 'loading',
-  username: string,
-} | {
-  mode: 'nothingSelected'
-}
-
-class UserPane extends Component<void, Props, void> {
-  render () {
-    if (this.props.mode === 'keybase') {
-      return <UserInfo {...this.props.userInfoProps} />
-    } else if (this.props.mode === 'external') {
-      return <NonUserInfo {...this.props.nonUserInfoProps} />
-    } else if (this.props.mode === 'loading') {
-      return <Loading username={this.props.username} />
-    }
-
-    return (
-      <Help />
-    )
-  }
-
-  static parseRoute () {
-    return {
-      componentAtTop: {title: 'Search'},
-    }
-  }
-}
+type OwnProps = { }
 
 const connector: TypedConnector<TypedState, TypedDispatch<SearchActions>, OwnProps, Props> = new TypedConnector()
 
 export default connector.connect(
   ({search: {userForInfoPane}, tracker: {trackers}}, dispatch, ownProps) => {
-    // Allow local overriding
-    if (ownProps.userForInfoPane) {
-      userForInfoPane = ownProps.userForInfoPane
-    }
-
     if (userForInfoPane && userForInfoPane.service === 'keybase') {
       const username = userForInfoPane.username
       const trackerState = trackers[username]
@@ -115,4 +64,4 @@ export default connector.connect(
         mode: 'nothingSelected',
       }
     }
-  })(UserPane)
+  })(Render)
